# internal/api/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: longue-vue CMDB API
  version: 0.1.0
  summary: Kubernetes-centric CMDB aligned with the ANSSI SecNumCloud framework.
  description: |
    REST API for **longue-vue**, a CMDB for Kubernetes environments aligned with the
    ANSSI SecNumCloud (SNC) qualification framework.

    This version (0.1.0) is a walking skeleton covering:
      - Liveness and readiness probes
      - `Cluster` resource CRUD

    Subsequent versions will add Node, Namespace, Workload, Pod, and other
    Kubernetes entity kinds mapped onto the ANSSI cartography layers.

    Errors follow RFC 7807 (`application/problem+json`).

    Interactive API documentation is available at `/docs/` for authenticated
    operators; sign in via `/ui/login` first to use "Try it out".
servers:
  - url: /
    description: Current origin (recommended for browser-based "Try it out" via /docs/)
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/kube-node-vms:
    get:
      operationId: listKubeNodeVMs
      summary: List kube-tagged cloud VMs and their reconciliation status
      description: |
        Cloud VMs carrying a Kubernetes cluster tag (OscK8sClusterID/* or
        OscK8sNodeName) reported by the VM collector, with the status
        computed at ingest (ADR-0045): `node`, `pending`, `orphan`,
        `unknown_cluster`. When `status` is absent the list defaults to
        `orphan` + `unknown_cluster`.
      tags: [virtual-machines]
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - { name: cloud_account_id, in: query, schema: { type: string, format: uuid } }
        - { name: cluster_id, in: query, schema: { type: string, format: uuid } }
        - name: status
          in: query
          description: Repeatable. Defaults to orphan and unknown_cluster.
          schema: { type: array, items: { type: string, enum: [node, pending, orphan, unknown_cluster] } }
          style: form
          explode: true
        - { name: cluster_hint, in: query, schema: { type: string } }
        - { name: instance_type, in: query, schema: { type: string } }
        - { name: power_state, in: query, schema: { type: string } }
        - name: name
          in: query
          description: Case-insensitive substring (or *-glob) on name and provider VM id.
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: sort
          in: query
          schema: { type: string, enum: [name, status, instance_type, power_state, status_since, first_seen_at, last_seen_at, provider_creation_date] }
        - $ref: "#/components/parameters/SortOrder"
      responses:
        "200":
          description: Kube-tagged VMs.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/KubeNodeVM" } }
                  next_cursor: { type: string, nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/kube-node-vms/summary:
    get:
      operationId: summarizeKubeNodeVMs
      summary: Per-account, per-status counts of kube-tagged VMs
      tags: [virtual-machines]
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - { name: cloud_account_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Summary rows.
          content:
            application/json:
              schema:
                type: object
                required: [rows]
                properties:
                  rows: { type: array, items: { $ref: "#/components/schemas/KubeNodeVMSummaryRow" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: |
        API token issued out-of-band. Send as `Authorization: Bearer <token>`.

        Scopes granted to a token:
          - `read`   — list and get operations
          - `write`  — create and update (implies `read` at the caller's discretion)
          - `delete` — remove resources
          - `admin`  — all of the above; reserved for future admin-only endpoints

        Tokens carrying the `admin` scope implicitly satisfy any scope required
        by an operation.

        Per ADR-0007, machine tokens are created by an admin in the UI
        (see `POST /v1/admin/tokens`) rather than injected through env vars.

    SessionCookie:
      type: apiKey
      in: cookie
      name: longue_vue_session
      description: |
        Human session cookie set by `POST /v1/auth/login`. HttpOnly,
        SameSite=Strict, 8-hour sliding expiry. Per ADR-0007.

        Every operation accepts either `BearerAuth` or `SessionCookie` —
        scope enforcement reads from the caller's role (for sessions) or
        token scopes (for bearer), so the two paths share the same
        downstream authorisation check.
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return. Server clamps to [1, 200].
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned from a previous list response.
      schema:
        type: string
        maxLength: 512
...
    SortOrder:
      name: order
      in: query
      required: false
      description: |
        Sort direction. Defaults to `asc` when `sort` is set;
        ignored when `sort` is absent.
      schema:
        type: string
        enum: [asc, desc]
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Token is valid but lacks the scope required by the operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      description: RFC 7807 problem details.
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    KubeNodeVM:
      type: object
      required: [id, cloud_account_id, cloud_account_name, provider_vm_id, name, status, status_since, first_seen_at, last_seen_at]
      properties:
        id: { type: string, format: uuid }
        cloud_account_id: { type: string, format: uuid }
        cloud_account_name: { type: string }
        provider_vm_id: { type: string }
        cluster_tag: { type: string }
        node_name_tag: { type: string }
        cluster_hint: { type: string, description: "Cluster name derived from the CAPO machine name" }
        name: { type: string }
        instance_type: { type: string }
        power_state: { type: string }
        zone: { type: string }
        vpc_id: { type: string }
        image_id: { type: string }
        image_name: { type: string }
        provider_creation_date: { type: string, format: date-time, nullable: true }
        node_id: { type: string, format: uuid, nullable: true }
        cluster_id: { type: string, format: uuid, nullable: true }
        cluster_name: { type: string, nullable: true }
        status: { type: string, enum: [node, pending, orphan, unknown_cluster] }
        status_since: { type: string, format: date-time }
        first_seen_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
    KubeNodeVMSummaryRow:
      type: object
      required: [cloud_account_id, cloud_account_name, status, count, vcpu, memory_gib]
      properties:
        cloud_account_id: { type: string, format: uuid }
        cloud_account_name: { type: string }
        status: { type: string, enum: [node, pending, orphan, unknown_cluster] }
        count: { type: integer }
        vcpu: { type: integer, description: "Sum of vCPU parsed from instance types (0 for unparsable types)" }
        memory_gib: { type: integer }